grc_backend/tprm_backend/requirements.txt
# Core Django packages
Django>=4.2.7,<6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.4
drf-spectacular==0.26.0
drf-yasg==1.21.7

# Database packages
PyMySQL==1.1.0
mysqlclient==2.2.0
mysql-connector-python>=8.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Security packages
django-environ==0.11.2
cryptography==41.0.7
bcrypt==4.1.2
django-ratelimit==4.1.0
django-defender==0.9.7
PyJWT==2.8.0

# Authentication & Authorization
django-oauth-toolkit==1.7.1

# Task Queue & Caching
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-redis==5.4.0
django-cacheops==7.2

# Input validation
cerberus==1.3.5

# File Processing & OCR
Pillow==10.1.0
pytesseract==0.3.10
PyMuPDF==1.23.8
openpyxl==3.1.2

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Storage & AWS
django-storages==1.14.2
boto3==1.34.0

# Async & HTTP
aiohttp==3.9.1
asyncio==3.4.3
requests==2.31.0

# Logging & Monitoring
django-extensions==3.2.3
python-json-logger==2.0.7
django-debug-toolbar==4.2.0

# Utilities
python-decouple==3.8
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz>=2023.3

# Import/Export & History
django-import-export==3.3.4
django-cleanup==8.1.0
django-simple-history==3.4.0
django-constance==2.9.1
django-picklefield==3.3

# Backup & Recovery
django-dbbackup==4.0.2

# Development & Testing
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0
